# Updates quantile bins to use native PostgreSQL precentile functions