--- a/NEWS.md
+++ b/NEWS.md
@@ -1,4 +1,4 @@
-0.21.1 (2018-03-08)
+0.22.0 (TBD)
 * Refactors CDB_QuantileBins to rely on PostgreSQL function `percentile_disc` #316
 
 0.21.0 (2018-02-15)