--- a/test/CDB_QuantileBinsTest.sql
+++ b/test/CDB_QuantileBinsTest.sql
@@ -6,11 +6,11 @@ SELECT unnest(CDB_QuantileBins(s, 10))
   FROM data;
 
 WITH data_nulls AS (
-    SELECT array_agg(x::numeric)
+    SELECT array_agg(x::numeric) AS s
       FROM (
         SELECT x FROM generate_series(0, 99) AS x
         UNION ALL
-        SELECT null AS s FROM generate_series(1, 10) AS x
+        SELECT null AS x FROM generate_series(1, 10) AS x
         ) _wrap
     )
 SELECT unnest(CDB_QuantileBins(s, 10))