This PR updates the quantiles function to be based firmly on PostgreSQL's function. https://www.postgresql.org/docs/9.6/static/functions-aggregate.html

# TODO

- [x] get function marked as PARALLEL

cc @stuartlynn