@ethervoid 2018-03-08
@andy-esch one more thing, could you please add a new entry in the NEWS.md with the information of this change :)