@andy-esch 2018-03-08
Ok, after a couple of dumb mistakes on my part, this is updated with @ethervoid's suggested changes. Could you give it one more look?

And is there anyone else who should take a look at this since it touches so much of our stack (Builder styling, TurboCarto, etc.). 

NEWS.md updates coming 👍