@ethervoid 2018-03-08
@andy-esch I've asked @jgoizueta to take a look at it and we should test it extensively in staging to avoid breaking a lot of functionality