@jgoizueta 2018-03-08
Hey, this looks very nice. Apart from the correctness, have we noticed any speed gains with the new implementation?