@ethervoid 2018-03-08
Taking this from the Jupiter notebook @andy-esch  did `The time difference in running this function off of 3000 values is less than 1/100th of a second on average.`

I think it worths gather new metrics using the last changes, immutable and using the array version of the `percentile_disc` function, to see if we have speed improvements too