@andy-esch 2018-03-08
By copying and pasting the function into my account, this is how the performance compares off of 2000 runs:
![quantiles_performance](https://user-images.githubusercontent.com/1041056/37164259-c9767c06-22c7-11e8-8011-03f1335590df.png)

script to produce that:
```python
%matplotlib inline
import matplotlib.pyplot as plt
import numpy as np
import cartoframes
from random import random
import time
from tqdm import tqdm

cc = cartoframes.CartoContext()
used_random = set()
times_new = []
times_old = []

for i in tqdm(range(2000)):
    r = random()
    while r in used_random:
        r = random()
    used_random.add(r)
    t1 = time.time()
    cc.query('''
    select pae_quantilebins(array(select mag::numeric from all_month_3), 7), {} as r
    '''.format(r))
    times_new.append(time.time() - t1)
    t1 = time.time()
    cc.query('''
    select cdb_quantilebins(array(select mag::numeric from all_month_3), 7), {} as r
    '''.format(r))
    times_old.append(time.time() - t1)
    
plt.semilogy(times_new)
plt.semilogy(times_old)
plt.semilogy([np.mean(times_new)] * len(times_new))
plt.semilogy([np.mean(times_old)] * len(times_old))
plt.legend([
    'New',
    'Old',
    'New Avg: {:0.4f}'.format(np.mean(times_new)),
    'Old Avg: {:0.4f}'.format(np.mean(times_old))]);
plt.ylabel('Response time (s)')
```