@andy-esch 2018-03-12
NEWS has been updated :) Is there anything else that needs to be done? @jgoizueta any other comments?