@ethervoid 2018-03-12
@rochoa could you please take a quick look and see if you see any problem changing this function ( as turbo-carto creator :) )