@ethervoid 2018-03-12
### QuatileBins acceptance

This is a extensively used function so we have to be extremly cautious with this change. I've been searching and checking where is used to compile a list of tests / steps to follow in order to check for possible problems.

#### Steps

- [ ] Create a new version to test it in staging and deploy it. We have to follow [this](http://doc-internal.cartodb.net/platform/internal_architecture/services/postgresql/extensions/deploy_cartodb_postgresql_extension.html?highlight=extension) guide
- [ ] Create a new branch for cartodb and [point the `lib/sql` submodule](http://doc-internal.cartodb.net/platform/internal_architecture/services/postgresql/extensions/deploy_cartodb_postgresql_extension.html?highlight=extension#deploy-and-test-with-cartodb-cartodb) to the new version commit (we should no create a new branch yet)
- [ ] Pass all the acceptance steps

#### Use of the function

I've been checking where the QuantileBins and JenksBins are used, the former uses the latter internally, in order to narrow a bit the kind of tests we have to perform in order to do the acceptance

* Builder

  - We use it in the styling so when we create a new map, we can go to the style option in the layer options, change size/color by value and there we can select the quantification method.
  - It's also used for the `describeFloat` function which is used by Pecan but AFAIK is not being used right now

* Turbo-Carto

  - Used to create the ramps and is linked with the first part of the `Builder` points


#### Acceptance steps (waiting for dataset samples to use in the tests)

1. Create a map with different datasets to test the multiple options for Quantile and Jenks bins
2. Use the builder to style the layers/maps using quantile and jenks bins
3. Check the console for errors
4. Check for consistency in the bucket generation