@rochoa 2018-03-12
I guess it's OK as far as the signature is the same —and it is.

I was wondering how good the performance improvement is, I tested it against a ~850k points dataset:

```
-- Current implementation
EXPLAIN ANALYZE
select cdb_quantilebins((select array_agg(temp) from (select * from ow) _q), 5);

                                                        QUERY PLAN                                                         
---------------------------------------------------------------------------------------------------------------------------
 Result  (cost=27115.34..27115.60 rows=1 width=32) (actual time=1989.101..1989.103 rows=1 loops=1)
   InitPlan 1 (returns $0)
     ->  Aggregate  (cost=27115.33..27115.34 rows=1 width=32) (actual time=555.857..555.859 rows=1 loops=1)
           ->  Seq Scan on ow  (cost=0.00..24960.86 rows=861786 width=32) (actual time=0.036..253.825 rows=861786 loops=1)
 Planning time: 0.740 ms
 Execution time: 2003.961 ms
(6 rows)

-- New implementation as proposed in this PR
EXPLAIN ANALYZE
select percentile_disc('{.2,.4,.6,.8,1}'::numeric[]) within group (order by temp asc) as p
from unnest((select array_agg(temp) from (select * from ow) _q)) as temp;

                                                         QUERY PLAN                                                         
----------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=27116.59..27116.60 rows=1 width=32) (actual time=1327.138..1327.138 rows=1 loops=1)
   InitPlan 1 (returns $0)
     ->  Aggregate  (cost=27115.33..27115.34 rows=1 width=32) (actual time=410.945..410.947 rows=1 loops=1)
           ->  Seq Scan on ow  (cost=0.00..24960.86 rows=861786 width=32) (actual time=0.010..137.112 rows=861786 loops=1)
   ->  Function Scan on unnest temp  (cost=0.00..1.00 rows=100 width=32) (actual time=560.324..700.954 rows=861786 loops=1)
 Planning time: 0.187 ms
 Execution time: 1346.070 ms
(7 rows)

-- The same approach but avoiding the array agg/unnest part:
EXPLAIN ANALYZE
select percentile_disc('{.2,.4,.6,.8,1}'::numeric[]) within group (order by temp asc) as p
from (select * from ow) _q;
                                                    QUERY PLAN                                                     
-------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=27115.33..27115.34 rows=1 width=32) (actual time=836.287..836.288 rows=1 loops=1)
   ->  Seq Scan on ow  (cost=0.00..24960.86 rows=861786 width=32) (actual time=0.009..148.319 rows=861786 loops=1)
 Planning time: 0.070 ms
 Execution time: 838.927 ms
(4 rows)
psql end.
```

It looks like performance is way better, but we are still penalized by the array aggregation/unnesting. I think it makes sense to have a new version that avoids/discourages the usage of the array aggregation/unnesting, and we start using it for our internal usages —e.g. Turbo-Carto. What do you think?